Move CategorySelector out of category components folder. Add prettier config. Add prettier and eslint compat. Add formatting commands. Format
