+++ frontend/components/category/ExplanationQuadrants.module.css
