+++ frontend/components/category/Quadrant.module.css
@@ -1,4 +1,3 @@
-
 .wrapper {
   background-color: var(--color-fg);
   border: 1px solid var(--color-green1);
@@ -12,7 +11,6 @@
     height: auto;
   }
 }
-
 
 @media (min-width: 768px) {
   .wrapper {
