+++ frontend/package.json
@@ -7,6 +7,8 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
+    "format": "prettier --write --ignore-path .gitignore .",
+    "format-check": "prettier --check --ignore-path .gitignore .",
     "generate-data": "ts-node --skip-project scripts/csvToJson.ts && prettier --write data/categories.ts"
   },
   "dependencies": {
@@ -21,6 +23,7 @@
     "@types/react-dom": "18.0.6",
     "eslint": "8.23.0",
     "eslint-config-next": "12.3.0",
+    "eslint-config-prettier": "^8.5.0",
     "prettier": "^2.7.1",
     "ts-node": "^10.9.1",
     "typescript": "4.8.3"
