+++ frontend/prettierrc.json
@@ -1,0 +1,6 @@
+{
+  "singleQuote": "false",
+  "tabWidth": 2,
+  "semi": true,
+  "trailingComma": "all"
+}
