+++ frontend/styles/Home.module.css
@@ -12,7 +12,7 @@
   border-bottom: 1px solid var(--color-purple);
   line-height: 1.5;
   font-size: 3.6rem;
-  width: 100%; 
+  width: 100%;
 }
 
 .description {
